ajustes nas funções
